pyproject.toml
[project]
name = "avellaneda-stoikov-mm"
version = "0.1.0"
description = "Lean Avellaneda-Stoikov Market Maker for Lighter Exchange"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]
monitoring = [
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
]

[tool.uv]
dev-dependencies = []

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]  # Line too long - handled by black

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"